Refactor: outsource proposal routes into proposals.ts
